Fixed a notice on certain situations
the custom_page_query "thinks" that $query is the query for the current page - BUT it's also called for other queries from other plugins. Hence get_queried_object is not guaranteed to return a term.

Fixed by adding a new is_main_query() condition. is_main_query() should only return true for the query corresponding to the current page.

Not fully tested.